Update microblog structure with blueprints.

+++ app/cli.py
@@ -1,44 +1,41 @@
-from app import app
 import click
 import os
 
 
-@app.cli.group()
-def translate():
-    """Translation and location commands."""
-    pass
+def register(app):
+    @app.cli.group()
+    def translate():
+        """Translation and location commands."""
+        pass
 
+    @translate.command()
+    @click.argument('lang')
+    def init(lang):
+        """Initialize a new language."""
+        extract = 'pybabel extract -F babel.cfg -k _l -o messages.pot .'
+        init = f'pybabel init -i messages.pot -d app/translations -l {lang}'
 
-@translate.command()
-@click.argument('lang')
-def init(lang):
-    """Initialize a new language."""
-    extract = 'pybabel extract -F babel.cfg -k _l -o messages.pot .'
-    init = f'pybabel init -i messages.pot -d app/translations -l {lang}'
+        if os.system(extract):
+            raise RuntimeError('extract command failed')
 
-    if os.system(extract):
-        raise RuntimeError('extract command failed')
+        if os.system(init):
+            raise RuntimeError('init command failed')
 
-    if os.system(init):
-        raise RuntimeError('init command failed')
+        os.remove('messages.pot')
 
-    os.remove('messages.pot')
+    @translate.command()
+    def update():
+        """Update all languages."""
+        if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
+            raise RuntimeError('extract command failed')
 
+        if os.system('pybabel update -i messages.pot -d app/translations'):
+            raise RuntimeError('update command failed')
 
-@translate.command()
-def update():
-    """Update all languages."""
-    if os.system('pybabel extract -F babel.cfg -k _l -o messages.pot .'):
-        raise RuntimeError('extract command failed')
+        os.remove('messages.pot')
 
-    if os.system('pybabel update -i messages.pot -d app/translations'):
-        raise RuntimeError('update command failed')
-
-    os.remove('messages.pot')
-
-
-@translate.command()
-def compile():
-    """Compile all languages."""
-    if os.system('pybabel compile -d app/translations'):
-        raise RuntimeError('compile command failed')
+    @translate.command()
+    def compile():
+        """Compile all languages."""
+        if os.system('pybabel compile -d app/translations'):
+            raise RuntimeError('compile command failed')
